before merge

+++ src/main/java/com/pigeon_stargram/sns_clone/service/post/PostLikeCrudService.java
@@ -1,7 +1,7 @@
 package com.pigeon_stargram.sns_clone.service.post;
 
-import com.pigeon_stargram.sns_clone.config.redis.ReadCacheKeyGenerator;
 import com.pigeon_stargram.sns_clone.domain.post.PostLike;
+import com.pigeon_stargram.sns_clone.domain.user.User;
 import com.pigeon_stargram.sns_clone.repository.post.PostLikeRepository;
 import com.pigeon_stargram.sns_clone.service.redis.RedisService;
 import lombok.RequiredArgsConstructor;
@@ -11,6 +11,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
+
+import static com.pigeon_stargram.sns_clone.constant.CacheConstants.*;
+import static com.pigeon_stargram.sns_clone.util.RedisUtil.*;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -23,8 +26,6 @@ public class PostLikeCrudService {
     private final PostLikeRepository repository;
     private final PostLikeRepository postLikeRepository;
 
-    private final ReadCacheKeyGenerator cacheKeyGenerator;
-
     public Optional<PostLike> findByUserIdAndPostId(Long userId,
                                                     Long postId) {
         return postLikeRepository.findByUserIdAndPostId(userId, postId);
@@ -34,14 +35,29 @@ public class PostLikeCrudService {
         return postLikeRepository.save(postLike);
     }
 
+    public void saveInCache(PostLike postLike) {
+        Long postId = postLike.getPost().getId();
+        Long userId = postLike.getUser().getId();
+        String cacheKey = cacheKeyGenerator(POST_LIKE_USER_IDS, POST_ID, postId.toString());
+
+        redisService.addToSet(cacheKey, userId);
+    }
+
     public void delete(PostLike postLike) {
         postLikeRepository.delete(postLike);
     }
 
+    public void deleteInCache(PostLike postLike) {
+        Long postId = postLike.getPost().getId();
+        Long userId = postLike.getUser().getId();
+        String cacheKey = cacheKeyGenerator(POST_LIKE_USER_IDS, POST_ID, postId.toString());
+
+        redisService.removeFromSet(cacheKey, userId);
+    }
+
     public Integer countByPostId(Long postId) {
         // 수동 캐시
-        String cacheKey = (String) cacheKeyGenerator
-                .generate("userId", "countByPostId", List.of(postId));
+        String cacheKey = cacheKeyGenerator(POST_LIKE_USER_IDS, POST_ID, postId.toString());
 
         if (redisService.hasKey(cacheKey)) {
             log.info("countByPostId {} 캐시 히트", postId);
@@ -49,15 +65,20 @@ public class PostLikeCrudService {
         }
         log.info("countByPostId {} 캐시 미스", postId);
 
-        // 좋아요 0개일때에도 캐시하기위해 빈 자료구조 생성
-        redisService.addToSet(cacheKey, "dummy");
-        redisService.removeFromSet(cacheKey, "dummy");
-
         // DB 조회후 레디스에 캐시
         List<PostLike> postLikes = repository.findByPostId(postId);
-        postLikes.forEach(postLike -> {
-            redisService.addToSet(cacheKey, postLike.getUser().getId());
-        });
+
+        if (postLikes.isEmpty()) {
+            // 좋아요 0개일때에도 캐시하기위해 빈 자료구조 생성
+            redisService.addToSet(cacheKey, "dummy");
+            redisService.removeFromSet(cacheKey, "dummy");
+        } else {
+            List<Long> postLikeUserIds = postLikes.stream()
+                    .map(PostLike::getUser)
+                    .map(User::getId)
+                    .toList();
+            redisService.addAllToSet(cacheKey, postLikeUserIds);
+        }
 
         return postLikes.size();
     }
